requirements.txt
aiogram~=3.0.0b7
requests~=2.31.0
beautifulsoup4~=4.12.2
mysql-connector-python~=8.0.33
emoji~=2.6.0